refactor: RoomParticipantRepository로 QueryDsl 변경 및 테스트 코드 수정
