+++ src/main/java/com/travel/role/domain/room/repository/querydsl/RoomParticipantQuerydsl.java
@@ -1,0 +1,5 @@
+package com.travel.role.domain.room.repository.querydsl;
+
+public interface RoomParticipantQuerydsl {
+	boolean existsUserInRoom(String email, Long roomId);
+}
