+++ src/main/java/com/travel/role/domain/room/repository/querydsl/RoomQuerydsl.java
@@ -7,5 +7,4 @@ import com.querydsl.core.Tuple;
 public interface RoomQuerydsl {
 
 	List<Tuple> getMemberInRoom(String email);
-	boolean existsUserInRoom(String email, Long roomId);
 }
