+++ src/main/java/com/travel/role/domain/room/repository/querydsl/RoomQuerydslImpl.java
@@ -39,13 +39,4 @@ public class RoomQuerydslImpl implements RoomQuerydsl {
 					.where(user.email.eq(email))
 			)).fetch();
 	}
-
-	public boolean existsUserInRoom(String email, Long roomId) {
-		Integer fetchOne = queryFactory
-			.selectOne()
-			.from(roomParticipant)
-			.where(roomParticipant.user.email.eq(email), roomParticipant.room.id.eq(roomId))
-			.fetchFirst();
-		return fetchOne != null;
-	}
 }
